+++ tests/test_utils.py
@@ -13,7 +13,7 @@ from rio_tiler import colormap, utils
 from rio_tiler.constants import WEB_MERCATOR_TMS, WGS84_CRS
 from rio_tiler.errors import RioTilerError
 from rio_tiler.expression import parse_expression
-from rio_tiler.io import COGReader
+from rio_tiler.io import Reader
 
 from .conftest import requires_webp
 
@@ -281,7 +281,7 @@ def test_cutline():
 
     feature_bounds = featureBounds(feat)
 
-    with COGReader(COGEO) as cog:
+    with Reader(COGEO) as cog:
         cutline = utils.create_cutline(cog.dataset, feat, geometry_crs="epsg:4326")
         data, mask = cog.part(feature_bounds, vrt_options={"cutline": cutline})
         assert not mask.all()
@@ -306,7 +306,7 @@ def test_cutline():
         },
     }
 
-    with COGReader(COGEO) as cog:
+    with Reader(COGEO) as cog:
         with pytest.raises(RioTilerError):
             utils.create_cutline(cog.dataset, feat_line, geometry_crs="epsg:4326")
 
@@ -334,7 +334,7 @@ def test_cutline():
         ],
     }
 
-    with COGReader(COGEO) as cog:
+    with Reader(COGEO) as cog:
         c = utils.create_cutline(cog.dataset, feat_mp, geometry_crs="epsg:4326")
         assert "MULTIPOLYGON" in c
 
@@ -353,7 +353,7 @@ def test_cutline():
         ],
     }
 
-    with COGReader(COGEO) as cog:
+    with Reader(COGEO) as cog:
         with pytest.raises(RioTilerError):
             utils.create_cutline(cog.dataset, bad_poly, geometry_crs="epsg:4326")